# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notifications API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "jamshedzodnekruz@gmail.com"
        },
        "version": "1.0"
    },
    "host": "api-notifications.dev.my.cloud",
    "basePath": "/api",
    "paths": {
...
        "/notifications-internal/v1/events": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get list of events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "apply filter with id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "apply filter with status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "apply filter with topic",
                        "name": "topic",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "apply filter with limit, 10 settled by default",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "apply filter with offset, 0 settled by default",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/event.eventModel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Create event",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/event.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/event.eventModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/notifications-internal/v1/events/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Update event",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/event.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/event.eventModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Delete event",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/notifications-internal/v1/events/{id}/image/{language}": {
            "post": {
                "description": "Provider event id - `{id}` and multi-lang image key - `{language}` as api route var to upload image for event",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Upload image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Load images with png, jpg, jpeg extensions",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/event.eventModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Provider event id - `{id}` and multi-lang image key - `{language}` as api route var to remove specific image of event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Remove image",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/event.eventModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "event.eventModel": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/language.Language"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "extraData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "$ref": "#/definitions/language.Language"
                },
                "link": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/language.Language"
                },
                "topic": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "event.request": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/language.Language"
                },
                "category": {
                    "type": "string"
                },
                "extraData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "link": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/language.Language"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "language.Language": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "ru": {
                    "type": "string"
                },
                "tg": {
                    "type": "string"
                },
                "uz": {
                    "type": "string"
                }
            }
        },
...
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "payload": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}